requirements.txt
fastapi
uvicorn[standard]
python-multipart
jinja2
opencv-python
mediapipe
numpy
pillow
scikit-learn
joblib
tqdm
pandas
